Combining View using stacks
To combine stacks, press control and click on the view which needs to be converted into vertical or horizontal stack, by selecting Embed in VStack or HStack.
We can add new element by clicking + icon on top right.
Place the items inside the VStack and Hstack according to design.
For adding Space in between use Spacer()